minor edits

+++ code/ThinTiles.py
@@ -11,9 +11,11 @@ import glob
 import numpy as np
 
 DataFolder = '/media/patrice/DataDrive/SEE_ICE/Train/' 
+RemainingTiles=40000
 
 for f in range(1,8):
     ilist=glob.glob(DataFolder+'C'+str(f)+'/*.tif')
-    idx = np.random.choice(np.arange(len(ilist)), len(ilist)-75000, replace=False)
-    for i in range(len(idx)):
-        os.remove(ilist[idx[i]])
+    if len(ilist)>RemainingTiles:
+        idx = np.random.choice(np.arange(len(ilist)), len(ilist)-RemainingTiles, replace=False)
+        for i in range(len(idx)):
+            os.remove(ilist[idx[i]])
